+++ tests/sample_data/data_idps.py
@@ -1,0 +1,13 @@
+from datetime import datetime
+
+data_idps = [
+    dict(
+        resource_hdx_id="90deb235-1bf5-4bae-b231-3393222c2d01",
+        admin2_ref=2,
+        assessment_type="BA",
+        reporting_round=18,
+        population=25000,
+        reference_period_start=datetime(2024, 1, 1),
+        reference_period_end=datetime(2024, 12, 31),
+    ),
+]