+++ changelog.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.8.16 (in progress)
+## 0.8.16
 
 ### Added
 
-- `data\_availability\_vat`
+- `idps` table, view, and VAT
+- `returnees` table, view, and VAT
+- `data_availability` VAT
+- new constraint: `greater\_than\_constraint`
 
 ### Changed
 
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added idps table and view (including materialised view)
 - Applied feedback from this
   [SQLAlchemy discussion](sqlalchemy/sqlalchemy#11748)
   to allow usage of type annotations defined in Base classes